frontdoor: moving the customizeDiff out
