+++ azurerm/internal/services/frontdoor/frontdoor_resource.go
@@ -472,13 +472,7 @@ func resourceArmFrontDoor() *schema.Resource {
 			"tags": tags.Schema(),
 		},
 
-		CustomizeDiff: func(d *schema.ResourceDiff, v interface{}) error {
-			if err := validate.FrontdoorSettings(d); err != nil {
-				return fmt.Errorf("creating Front Door %q (Resource Group %q): %+v", d.Get("name").(string), d.Get("resource_group_name").(string), err)
-			}
-
-			return nil
-		},
+		CustomizeDiff: frontDoorCustomizeDiff,
 	}
 }
